+++ spring/hello-spring/hello-spring/src/test/java/hello/hellospring/HelloSpringApplicationTests.java
@@ -1,4 +1,4 @@
-package hello.hello_spring;
+package hello.hellospring;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
