+++ spring/hello-spring/hello-spring/src/test/java/hello/hellospring/service/MemberServiceTest.java
@@ -1,8 +1,7 @@
-package hello.hello_spring.service;
+package hello.hellospring.service;
 
 import hello.hellospring.domain.Member;
 import hello.hellospring.repository.MemoryMemberRepository;
-import hello.hellospring.service.MemberService;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
