+++ spring/hello-spring/hello-spring/src/test/java/hello/hellospring/service/SpringConfig.java
@@ -1,8 +1,7 @@
-package hello.hello_spring.service;
+package hello.hellospring.service;
 
 import hello.hellospring.repository.MemberRepository;
 import hello.hellospring.repository.MemoryMemberRepository;
-import hello.hellospring.service.MemberService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
